+++ deid/dicom/parser.py
@@ -79,7 +79,7 @@ class DicomParser:
     def __repr__(self):
         return self.__str__()
 
-    def load(self, dicom_file, force=True):
+    def load(self, dicom_file, force=True, stop_before_pixels=False):
         """Ensure that the dicom file exists, and use full path. Here
         we load the file, and save the dicom, dicom_file, and dicom_name.
         """
@@ -94,7 +94,9 @@ class DicomParser:
             # If we must read the file, the path must exist
             if not os.path.exists(dicom_file):
                 bot.exit("%s does not exist." % dicom_file)
-            self.dicom = read_file(dicom_file, force=force)
+            self.dicom = read_file(
+                dicom_file, force=force, stop_before_pixels=stop_before_pixels
+            )
 
         # Set class variables that might be helpful later
         self.dicom_file = os.path.abspath(self.dicom.filename)